+++ examples/2d_rendering/sources/example_2d_scene.m
@@ -6,6 +6,7 @@
 #include <metil_player/metil_player.h>
 #include <metil_positioning.h>
 #include <metil_rendering/metil_renderable.h>
+#include <metil_rendering/metil_renderable_type.h>
 #include <metil_rendering/metil_renderer_data_object.h>
 #include <metil_scenes/metil_scene.h>
