+++ examples/fog/sources/example_fog_scene.m
@@ -13,6 +13,7 @@
 #include <metil_object.h>
 #include <metil_player/metil_player.h>
 #include <metil_rendering/metil_renderable.h>
+#include <metil_rendering/metil_renderable_type.h>
 #include <metil_rendering/metil_renderer_data_object.h>
 #include <metil_scenes/metil_scene.h>
