+++ include/metil_rendering/metil_renderable_type.h
@@ -1,0 +1,11 @@
+#ifndef __metil_rendering_metil_renderable_type_h
+#define __metil_rendering_metil_renderable_type_h
+
+enum metil_renderable_type {
+  metil_renderable_type_group = 0,
+  metil_renderable_type_object = 1,
+  metil_renderable_type_menu = 2,
+  metil_renderable_type_model = 3
+};
+
+#endif
